requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.08.22=h06a4308_0
expat=2.5.0=h6a678d5_0
joblib=1.3.2=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
ncurses=6.4=h6a678d5_0
openssl=1.1.1w=h7f8727e_0
pillow=10.1.0=pypi_0
pip=23.3=py310h06a4308_0
pyparsing=3.1.1=pypi_0
python=3.10.0=h12debd9_5
python-dateutil=2.8.2=pypi_0
pyyaml=6.0.1=pypi_0
readline=8.2=h5eee18b_0
scikit-learn=1.3.2=pypi_0
setuptools=68.0.0=py310h06a4308_0
six=1.16.0=pypi_0
sqlite=3.41.2=h5eee18b_0
threadpoolctl=3.2.0=pypi_0
tk=8.6.12=h1ccaba5_0
tzdata=2023c=h04d1e81_0
wheel=0.41.2=py310h06a4308_0
xz=5.4.2=h5eee18b_0
zipp=3.17.0=pypi_0
zlib=1.2.13=h5eee18b_0
absl-py==2.0.0
chex==0.1.84
contourpy==1.2.0
cycler==0.12.1
etils==1.5.2
flax==0.7.5
fonttools==4.44.3
fsspec==2023.10.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
grpcio==1.59.2
importlib-resources==6.1.1
jax==0.4.19
jaxlib==0.4.19
joblib==1.3.2
kiwisolver==1.4.5
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.1
mdurl==0.1.2
ml-dtypes==0.3.1
msgpack==1.0.7
nest-asyncio==1.5.8
numpy==1.26.1
opt-einsum==3.3.0
optax==0.1.7
orbax-checkpoint==0.4.2
packaging==23.2
Pillow==10.1.0
protobuf==4.23.4
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
PyYAML==6.0.1
requests-oauthlib==1.3.1
rich==13.7.0
scikit-learn==1.3.2
scipy==1.11.3
six==1.16.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorstore==0.1.50
threadpoolctl==3.2.0
toolz==0.12.0
torch==1.11.0
torchtext==0.12.0
tqdm==4.66.1
typing_extensions==4.8.0
Werkzeug==3.0.1
zipp==3.17.0
